Allow the dataUrl to be specified in query params.
Note: this does not use the new query-params feature of Ember, it is
simply using the normal javascript query param parsing so that when
the application is booted, we pass it the correct dataUrl.

So navigating to the site with a url like:

/?dataUrl=/tests/support/api.json

Will do an XHR and use `/tests/support/api.json` for its data source.